requirements.txt
Pillow
requests
keyboard
pyttsx3
openai
